% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igstrd.R
\name{calc.zlen}
\alias{calc.zlen}
\title{calc.zlen}
\usage{
calc.zlen(mat, lenanthro)
}
\description{
Function for calculating individual Length-for-age z-scores
}
\details{
WHO Child Growth Standards Department of Nutrition for Health and Development Last modified on 07/10/2013-Developed using R version 3.0.1
}
\note{
This code conrcerns the standard approach for the prevalences, i.e. the calculation of the prevalences takes into account all the valid (non-missing) z-scores for each of the indicators.
}
\examples{
data("igrowupstd_AnXprs")
mat<-data.frame(age.days=c(1854,1854),sex=c(1,2),clenhei=c(110.4611,109.6604))
calc.zlen(mat,lenanthro)
}
\keyword{igrowup}
\keyword{who}
\keyword{zscore}
