% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igstrd.R
\name{igstd}
\alias{igstd}
\title{Modified igrowup}
\usage{
igstd(mydf, sex, age, age.month = F, weight = rep(NA, dim(mydf)[1]),
  lenhei = rep(NA, dim(mydf)[1]), measure = rep(NA, dim(mydf)[1]),
  headc = rep(NA, dim(mydf)[1]), armc = rep(NA, dim(mydf)[1]),
  triskin = rep(NA, dim(mydf)[1]), subskin = rep(NA, dim(mydf)[1]),
  oedema = rep("n", dim(mydf)[1]), sw = rep(1, dim(mydf)[1]))
}
\description{
Calculate the z-scores for the indicators: length/height-for-age, weight-for-age, weight-for-legnth/height and body mass index-for-age
}
\details{
WHO Child Growth Standards Department of Nutrition for Health and Development Last modified on 07/10/2013-Developed using R version 3.0.1
}
\note{
This code conrcerns the standard approach for the prevalences, i.e. the calculation of the prevalences takes into account all the valid (non-missing) z-scores for each of the indicators.
}
\keyword{igrowup}
\keyword{who}
\keyword{zscore}
