% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igstrd.R
\name{prevph}
\alias{prevph}
\title{Prevph}
\usage{
prevph(a, x, w, f)
}
\description{
Z-score prevalence calc With oedema (only for weight-for-length/height and bmi-for-age)
}
\details{
WHO Child Growth Standards Department of Nutrition for Health and Development Last modified on 07/10/2013-Developed using R version 3.0.1
}
\note{
This code conrcerns the standard approach for the prevalences, i.e. the calculation of the prevalences takes into account all the valid (non-missing) z-scores for each of the indicators.
}
\examples{
a<-c(1,1)
x<-c(1,1)
w<-c(1,1)
f<-"y"
prevph(a,x,w,f)
}
\keyword{igrowup}
\keyword{who}
\keyword{zscore}
