% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igstrd.R
\name{rounde}
\alias{rounde}
\title{Rounde}
\usage{
rounde(x, digits = 0)
}
\description{
Rounding function - SPlus rounding function uses the nearest even number rule
}
\details{
WHO Child Growth Standards Department of Nutrition for Health and Development Last modified on 07/10/2013-Developed using R version 3.0.1
}
\note{
This code conrcerns the standard approach for the prevalences, i.e. the calculation of the prevalences takes into account all the valid (non-missing) z-scores for each of the indicators.
}
\examples{
x<-c(1,1)
rounde(x)
}
\keyword{igrowup}
\keyword{who}
\keyword{zscore}
