% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igstrd.R
\name{wmean}
\alias{wmean}
\title{wmean}
\usage{
wmean(x, w)
}
\description{
Weighted mean
}
\details{
WHO Child Growth Standards Department of Nutrition for Health and Development Last modified on 07/10/2013-Developed using R version 3.0.1
}
\note{
This code conrcerns the standard approach for the prevalences, i.e. the calculation of the prevalences takes into account all the valid (non-missing) z-scores for each of the indicators.
}
\examples{
x<-c(1,1)
w<-c(1,1)
wmean(x,w)
}
\keyword{igrowup}
\keyword{who}
\keyword{zscore}
